+++ Assignment 3/Pop Alexia/Project/backend/src/main/java/com/lab4/demo/quizz/QuizzController.java
@@ -4,12 +4,10 @@ import com.lab4.demo.quizz.model.dto.QuizzDTO;
 import com.lab4.demo.security.dto.MessageResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.ConstraintViolationException;
 import java.util.List;
-import java.util.Set;
 
 import static com.lab4.demo.UrlMapping.QUIZZ;
 
@@ -31,40 +29,35 @@ public class QuizzController {
     @PostMapping
     public ResponseEntity<?> create( @RequestBody QuizzDTO quizzDTO) {
         if(quizzService.findByQuizzTitle(quizzDTO.getTitle())!=null){
-            throw new ConstraintViolationException("This quizz already exists", Set.of());
+            return ResponseEntity.badRequest().body(new MessageResponse("This quizz already exists"));
         }
-
-        return ResponseEntity.ok().body(quizzService.create(quizzDTO));
+        try {
+            return ResponseEntity.ok().body(quizzService.create(quizzDTO));
+        } catch (ConstraintViolationException e) {
+            return ResponseEntity.badRequest().body(new MessageResponse(e.getConstraintViolations().iterator().next().getMessage()));
+        }
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<?> edit(@PathVariable Long id ,  @RequestBody QuizzDTO quizzDTO) {
         if(!quizzService.findById(id).getTitle().equals(quizzDTO.getTitle())) {
             if(quizzService.findByQuizzTitle(quizzDTO.getTitle())!=null){
-                throw new ConstraintViolationException("This quizz already exists", Set.of());
+                return ResponseEntity.badRequest().body(new MessageResponse("This quizz already exists"));
             }
         }
-        quizzService.edit(id, quizzDTO);
-        return ResponseEntity.ok(new MessageResponse("Quizz edited successfully"));
+        try {
+            quizzService.edit(id, quizzDTO);
+            return ResponseEntity.ok(new MessageResponse("Quizz edited successfully"));
+        }catch (ConstraintViolationException e){
+            return ResponseEntity.badRequest().body(new MessageResponse(e.getConstraintViolations().iterator().next().getMessage()));
+        }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> delete(@PathVariable Long id) {
+        if(quizzService.findById(id)==null)
+            return ResponseEntity.badRequest().body(new MessageResponse("Quizz not found"));
         quizzService.delete(id);
         return ResponseEntity.ok(new MessageResponse("Quizz deleted successfully"));
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<?> handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
-        return ResponseEntity
-                .badRequest()
-                .body(new MessageResponse(e.getBindingResult().getFieldError().getDefaultMessage()));
-    }
-
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity<?> handleConstraintViolationException(ConstraintViolationException e) {
-        return ResponseEntity
-                .badRequest()
-                .body(new MessageResponse(e.getMessage()));
     }
 }
